fix: cardList 파일 경로 변경 및 불필요 파일 제거
