+++ src/components/card/ListPickerCardElement.jsx
@@ -1,6 +1,6 @@
 "use client";
 
-import PokemonCard from "@/components/cardList/list/card/PokemonCard";
+import PokemonCard from "@/components/card/PokemonCard";
 
 export default function CardElement({
   data,
